requirements.txt
langchain-openai>=0.1.0
python-dotenv>=1.0.0
langchain>=0.1.0
requests>=2.31.0
arcadepy>=0.1.0
tweepy>=4.14.0
supabase>=1.0.3
flask>=2.0.0
praw>=7.0.0
google-generativeai>=0.3.2
Pillow>=10.0.0
streamlit>=1.0.0
bcrypt>=3.2.0